requirements.txt
opencv-python
numpy
pymavlink
matplotlib
scipy
PyYAML
